pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pyapiclient"
version = "0.1.0"
description = "pyAPIClient: Django-like dynamic ORM models generated from OpenAPI 2/3 or GraphQL schemas"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "pyAPIClient contributors" }]
dependencies = [
  "httpx>=0.27,<1",
  "PyYAML>=6",
]

[project.optional-dependencies]
graphql = [
  "graphql-core>=3.2,<4",
]
dev = [
  "pytest>=8",
  "pytest-cov>=5",
  "respx>=0.21",
  "graphql-core>=3.2,<4",
  "pre-commit>=4",
]

[tool.hatch.build.targets.wheel]
packages = ["src/pyapiclient"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-q --cov=pyapiclient --cov-report=term-missing --cov-fail-under=90"

[tool.coverage.run]
branch = true
source_pkgs = ["pyapiclient"]

[tool.coverage.report]
exclude_lines = [
  "pragma: no cover",
  "if TYPE_CHECKING:",
]
